refactoring input stuff
