+++ src/main/scala/simulation/pcg/DungeonGenerator.scala
@@ -2,16 +2,17 @@ package simulation.pcg
 
 import game_object_system.Globals
 import game_object_system.graphics_objects.Sprite
+import javax.swing.text.html.parser.Entity
 
 object DungeonGenerator {
 
   /** Dungeon generated with a random width and height */
-  def generateDungeon() : List[Sprite] = {
-    val w = Globals.random.between(1000, 3001) // a inclusive b esclusive
-    val h = Globals.random.between(1000, 3001)
+  def generateDungeon() : List[Entity] = {
+    val w = Globals.random.between(300, 1001) // a inclusive b esclusive
+    val h = Globals.random.between(300, 1001)
     generateDungeon(w, h)
   }
 
-  def generateDungeon(w: Int, h : Int) : List[Sprite] = ???
+  def generateDungeon(w: Int, h : Int) : List[Entity] = ???
 
 }
